% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz_functions.r
\docType{data}
\name{corp}
\alias{corp}
\title{Corporatism}
\format{
A list with the following variables for 16 countries over 14 years:
\describe{
\item{y}{Economic growth}
\item{country}{Country indicator variable}
\item{imports}{Imports price movement}
\item{exports}{Export price movement}
\item{left}{Leftist government}
\item{demand}{Demand}
\item{growth.lag}{Lag of economic growth}
\item{labor.org}{Labor organizations prevalence, only one observation per country}
}
}
\source{
\url{https://spia.uga.edu/faculty_pages/rbakker/bayes/POLS\%20Bayes.htm}
}
\usage{
corp
}
\description{
A dataset from Alvarez, Garrett and Lange's 1991 APSR article.
}
\keyword{datasets}
