% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz_functions.r
\name{sigd3}
\alias{sigd3}
\title{Significance Plot with D3}
\usage{
sigd3(
  object,
  term,
  fname,
  height = 500,
  width = 625,
  return_iFrame = TRUE,
  ylab = paste0("Effect of ", term),
  axfont = 12,
  labfont = 12,
  labfam = c("sans-serif", "serif"),
  colors = c("red", "black", "black")
)
}
\arguments{
\item{object}{A model that has a categorical predictor.  The function
uses both \code{ggpredict} from the \pkg{ggeffects} package and
\code{factorplot} from the \pkg{factorplot} package (with the
variable given by the \code{factor.var=term} argument, so the object
must be amenable to these manipulations. )}

\item{term}{The term from the model to be plotted.}

\item{fname}{Name of the file to which the html code will be written.}

\item{height}{Height of the plot in pixels}

\item{width}{Width of the plot in pixels}

\item{return_iFrame}{Logical indicating whether an \code{iframe}
should be returned to hold the plot.  If \code{FALSE} only the html
file is produced and the function prpduces no other visible output.}

\item{ylab}{Label to put on y-axis.}

\item{axfont}{Font size for axis tick mark labels (in px).}

\item{labfont}{Font size for y-axis label (in px).}

\item{labfam}{Font-family for y-axis label (must be one of "serif" or "sans-serif").}

\item{colors}{Vector of three colors the first for the selected items, the second for the un-selected items and the third for when no points are hovered over.}
}
\value{
Either an html file or an iframe linked to the html file.
}
\description{
Makes an errorbar plot using the D3.js library with a hover-over
effect that highlights all cagtegories that are significantly differnet
than the one being hovered over.
}
