% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviz_functions.r
\name{sigd3h}
\alias{sigd3h}
\alias{sigd3h.default}
\alias{sigd3h.mcmc.list}
\alias{sigd3h.mcmc}
\title{Significance Plot with D3 (horizontal)}
\usage{
sigd3h(
  object,
  term,
  fname,
  height = 500,
  width = 625,
  return_iFrame = TRUE,
  level = 0.95,
  ylab = paste0("Effect of ", term),
  axfont = 12,
  labfont = 12,
  lmexpand = 0.15,
  labfam = c("sans-serif", "serif"),
  colors = c("#B80000", "#000000", "#000000"),
  ptSize = c(7, 5),
  lineSize = c(3, 1.5),
  names = NULL,
  order = c("size-descending", "size-ascending", "natural")
)
}
\arguments{
\item{object}{A model that has a categorical predictor.  The function
uses both \code{ggpredict} from the \pkg{ggeffects} package and
\code{factorplot} from the \pkg{factorplot} package (with the
variable given by the \code{factor.var=term} argument), so the object
must be amenable to these manipulations.  Alternatively, it could be
an mcmc object in which case all columns will be plotted.}

\item{term}{The term from the model to be plotted.}

\item{fname}{Name of the file to which the html code will be written.}

\item{height}{Height of the plot in pixels}

\item{width}{Width of the plot in pixels}

\item{return_iFrame}{Logical indicating whether an \code{iframe}
should be returned to hold the plot.  If \code{FALSE} only the html
file is produced and the function prpduces no other visible output.}

\item{level}{Confidence level at which to do the test (or in the mcmc case, credibility level for the credible intervals).}

\item{ylab}{Label to put on y-axis.}

\item{axfont}{Font size for axis tick mark labels (in px).}

\item{labfont}{Font size for y-axis label (in px).}

\item{lmexpand}{Factor by which the left-margin should be expanded to accommodate tick mark labels.}

\item{labfam}{Font-family for y-axis label (must be one of "serif" or "sans-serif").}

\item{colors}{Vector of three colors the first for the selected items, the second for the un-selected items and the third for when no points are hovered over.}

\item{ptSize}{Vector of two point sizes - the first for selected observations and the second for unselected ones.}

\item{lineSize}{Vector of line widths - the first for selected observations and the second for unselected ones.}

\item{names}{An optional vector of names for the parameters.}

\item{order}{How should parameters be organized by size-ascending, size-descending or natural (i.e., the order in the data).}
}
\value{
Either an html file or an iframe linked to the html file.
}
\description{
Makes an errorbar plot using the D3.js library with a hover-over
effect that highlights all cagtegories that are significantly differnet
than the one being hovered over. The orientation of this plot is
horizontal rather than vertical
}
